Feature: Add support for GCR buckets with uniform_bucket_level_access = true
The new default setting for buckets is to use bucket-level policies (instead of an ACL), this also applies to buckets created for the container-registry. These buckets don't support setting google_storage_bucket_access_control policies, so the ACL grant fails.

This PR uses the new (as of v3.88 of the google terraform provider) google_storage_bucket datasource to determine whether or not uniform_bucket_level_access = true is set, and if so uses the google_storage_bucket_iam_member resource to grant access.

Special thanks to @andyroyle and @patricklubach