feat: Change the release process (#60)
Use the new release-please-action GitHub action to release the Terraform
module. This new way to release the module will open a release pull
request that will append commits from main until we're ready to create
a new release. The new way avoids manual release from the Makefile.

Also, the release process now depends on Conventional Commits where this
convention frames the way the commit message should be written. The
messages must now contain a prefix that specifies the type of the
changes (fix, feature, etc...). The GitHub action will use this prefix
to sort the commit in the changelog. It will also use it to infer the
next version number. For example, if the release contains only fixes,
the new version will be a patch (ex: v1.0.0 -> v1.0.1).

Because the repository already has a tag and release, we have to include
some bootstrap configuration otherwise release-please will use 0.1.0 as
the next release instead of 1.3.1 or 1.4.0.

To make sure commit authors respect the new convention, this commit
introduces a new pre-commit hook that enforces it.